- AntSystem za pocetne tragove koristi pohlepni algoritam - Napravljen refactor Genetskog algoritma 	- dodani genericsi 	- za unutarnje predstavljanje jedinki se koristi java.util.List 	  zbog genericsa 	- populacija sadrzi RandomGenerator i Krajolik instance 	  (a ne Simulator)

+++ Projekt-EA/src/ea/aco/Grad.java
@@ -9,13 +9,13 @@ package ea.aco;
  */
 public class Grad {
 	
-	public final int x;
+	public final double x;
 	
-	public final int y;
+	public final double y;
 	
 	public final String ime;
 	
-	public Grad(int x, int y, String ime) {
+	public Grad(double x, double y, String ime) {
 		this.x = x;
 		this.y = y;
 		this.ime = ime;
